Added the ability to get private email addresses
A large percentage of github users don't have public email addresses.
Therefore, you need to request the scope: user:email or user to get the
email addresses. The issue is that the initial userinfo request doesn't
contain the private email addresses. This needs to be done in a separate
request
(https://developer.github.com/v3/users/emails/#list-email-addresses-for-a-user).

This code makes that separate request and grabs the first verified user.