Fixes #70 Unit tests failing
I believe this also fixes #73 as well.

@ProNotion , @knightmeister @titarenko Can you please review this
change. After looking at ParseTokenResponse I believe it should always
return a string, the reasoning behind this is all values can be
converted from a string (in one case there is an int) and it makes it
consistent if it's coming from a query string.. Previously this method
threw an exception if the key didn't exist. I don't agree with this as
someone might want to use this (as in the sales force client). Secondly
you might want to get optional token values out like expires, token type
and a refresh token.. I think the proper behavior would be to throw an
exception if something can't be parsed and it's required as I've done
with AccessToken. Thoughts?